device: mpu9250: update mpu9250 driver
update the mpu9250 driver with dmp mode
