--- device/peripheral/sensor/imu/mpu9250_dmp/mpu9250_dmp.mk
@@ -1,2 +1,0 @@
-DEV_CSRCDIR +=  $(EMBARC_ROOT)/device/peripheral/sensor/imu/mpu9250_dmp
-DEV_INCDIR +=  $(EMBARC_ROOT)/device/peripheral/sensor/imu/mpu9250_dmp
